# swagger.yaml
...
swagger: "2.0"
info:
  description: The API For Ops to Ocs communications
  version: 1.0.0
  title: owner-configuration-api
host: virtserver.swaggerhub.com
basePath: /rftemple/OwnerCompanion/1.0.0
schemes:
- https
paths:
...
  /v1/devices/{deviceId}/msgs:
    get:
      tags:
      - OCS
      summary: Retrieves all serviceInfo Messages to be sent to the device
      operationId: getServiceInfo
      produces:
      - application/json
      parameters:
      - name: deviceId
        in: path
        required: true
        type: string
      responses:
        200:
          description: response
          schema:
            $ref: '#/definitions/ArrayOfSviMessage'
        404:
          description: The device not found
        500:
          description: Server error
    post:
      tags:
      - OCS
      summary: Posts a deviceServiceInfo message recieved from the device
      operationId: postDeviceServiceInfo
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: deviceId
        in: path
        required: true
        type: string
      - in: body
        name: messages
        required: false
        schema:
          $ref: '#/definitions/ModuleMessage'
      responses:
        200:
          description: response
    put:
      tags:
      - Optional
      summary: Updates all service info messages to be sent to the device
      operationId: putServiceInfo
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: deviceId
        in: path
        required: true
        type: string
      - name: valueId
        in: path
        description: The Id of the ServiceInfo value to update
        required: true
        type: string
      - in: body
        name: messages
        required: false
        schema:
          $ref: '#/definitions/ModuleMessage'
      responses:
        200:
          description: response
...
definitions:
  ArrayOfSviMessage:
    type: array
    items:
      $ref: '#/definitions/SviMessage'
  SviMessage:
    type: object
    properties:
      module:
        type: string
        example: sdo_sys
        description: The name of the module sending the message
      msg:
        example: exec, filedesc, write, etc
        description: The name of the messenge being sent
      valueLen:
        type: integer
        description: the total length of the value of the message
      enc:
        type: string
        description: the intened encoding expected to use when sending the value
        enum:
        - base64
        - ascii
      valueId:
        type: string
        description: the id to use when retrieving the value of the message
...
  ModuleMessage:
    type: object
    properties:
      module:
       description: The module sending the message
       type: string
      msg:
        description: The name of the messenge being sent
        type: string
      value:
        type: string
        description: The value of the message
...